# internal/interfaces/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Authentication Service API
  description: |
    AuthService is a self-hosted, multi-tenant authentication platform for teams that want provider-grade auth without handing user data, pricing leverage, or service-to-service trust to a third party.

    It provides email/password auth, OAuth2 social login, enterprise SSO with SAML/OIDC, SCIM 2.0 directory sync, machine-to-machine client credentials, magic links, TOTP 2FA, adaptive security and step-up, passkeys/WebAuthn, organization RBAC, refresh-token rotation, tenant-scoped JWTs, JWKS, audit logs, REST, gRPC, and an importable Go JWT validator.

    ## What This Replaces

    Use AuthService when you would otherwise reach for Auth0, Clerk, Firebase Auth, Supabase Auth, Cognito, Stytch, WorkOS, or Keycloak but need a lightweight service your company can own, deploy, inspect, and extend.

    | Need | AuthService Capability |
    |------|------------------------|
    | Product authentication | Signup, login, refresh, logout, profile, password change |
    | Passwordless login | Magic links and passkeys/WebAuthn |
    | MFA | TOTP setup, enable, login verify, and disable lifecycle |
    | Adaptive security | Risk-driven MFA, remembered devices, admin/user step-up, and policy-driven challenge/block/notify decisions |
    | Social identity | Google, GitHub, Microsoft, and Apple OAuth2 |
    | Enterprise SSO | Per-client SAML 2.0 and OIDC connections, domain routing, metadata, and JIT provisioning |
...

...

...
servers:
...
paths:
...
  /api/admin/clients:
    post:
      operationId: createClient
      summary: Register a new client
      description: |
        Creates a new tenant (client) and returns the API key and JWT secret. **Save these values** — they are only shown at creation time.
      tags: [Admin]
      x-codeSamples:
        - lang: cURL
          label: Provision a production client
          source: |
            curl -X POST https://authservice.ayushojha.com/api/admin/clients \
              -H "X-Admin-Key: $AUTH_SERVICE_ADMIN_KEY" \
              -H "Content-Type: application/json" \
              -d '{
                "name": "Acme Dashboard",
                "slug": "acme-dashboard-prod",
                "allowed_origins": ["https://app.acme.com"],
                "webhook_url": "https://app.acme.com/webhooks/auth"
              }'
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
            examples:
              productionWebApp:
                summary: Production web application client
                value:
                  name: Acme Dashboard
                  slug: acme-dashboard-prod
                  allowed_origins:
                    - https://app.acme.com
                    - https://www.acme.com
                  webhook_url: https://app.acme.com/webhooks/auth
              stagingEnvironment:
                summary: Separate staging client for safer testing
                value:
                  name: Acme Dashboard Staging
                  slug: acme-dashboard-staging
                  allowed_origins:
                    - https://staging.acme.com
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClientResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listClients
      summary: List all clients
      description: Returns all registered clients.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/clients/{id}:
    get:
      operationId: getClient
      summary: Get a client by ID
      description: Returns a single client's details.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateClient
      summary: Update a client
      description: Updates mutable client settings such as allowed origins, webhook URL, status, and tenant WebAuthn policy controls.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
            example:
              settings:
                webauthn_attestation: direct
                webauthn_require_attestation: true
                webauthn_allowed_attestation_formats: [packed, tpm, apple]
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/clients/{id}/rotate-jwt:
    post:
      operationId: rotateJwtSecret
      summary: Rotate JWT secret (alias)
      description: |
        Alias for `/api/admin/clients/{id}/rotate-secret`. Generates a new JWT secret for the client. The old secret is immediately invalidated — all existing tokens signed with it will fail validation.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: New JWT secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateSecretResponse"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/clients/{id}/rotate-key:
    post:
      operationId: rotateApiKey
      summary: Rotate API key (alias)
      description: |
        Alias for `/api/admin/clients/{id}/rotate-api-key`. Generates a new API key for the client. The old key is immediately invalidated.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: New API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateKeyResponse"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Client API key — identifies the tenant
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token (JWT)
    SCIMBearerAuth:
      type: http
      scheme: bearer
      description: Per-directory SCIM provisioning token
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Break-glass master admin key for bootstrap or emergencies; prefer AdminBearerAuth
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token returned by `/api/admin/auth/login` or `/api/admin/auth/sso`
...
  schemas:
...
    CreateClientRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          description: Human-readable project name
          example: My SaaS App
        slug:
          type: string
          description: Unique identifier (URL-safe)
          example: my-saas-app
        allowed_origins:
          type: array
          items:
            type: string
          description: Allowed CORS origins
          example: ["https://myapp.com", "https://www.myapp.com"]
        webhook_url:
          type: string
          format: uri
          description: Webhook URL for signed audit-event deliveries when `WEBHOOK_SIGNING_SECRET` is configured
          example: https://myapp.com/webhooks/auth
        settings:
          type: object
          description: Tenant settings. WebAuthn policy keys include `webauthn_attestation`, `webauthn_require_attestation`, and `webauthn_allowed_attestation_formats`.
          additionalProperties: true
    UpdateClientRequest:
      type: object
      properties:
        name:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          format: uri
        status:
          type: string
          enum: [active, suspended]
        settings:
          type: object
          additionalProperties: true
          description: Replaces the client settings object when supplied.

    # ── Responses ─────────────────────────────
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: invalid request
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          nullable: true
          description: Receives signed `audit.event` webhooks when webhook delivery is enabled
        settings:
          type: object
        status:
          type: string
          enum: [active, suspended]
        token_mode:
          type: string
          enum: [v1_hs256, v2_jwks]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateClientResponse:
      type: object
      properties:
        client:
          $ref: "#/components/schemas/Client"
        api_key:
          type: string
          description: Client API key — save this, shown only at creation
        jwt_secret:
          type: string
          description: JWT signing secret — save this, shown only at creation
    RotateSecretResponse:
      type: object
      properties:
        client:
          $ref: "#/components/schemas/Client"
        jwt_secret:
          type: string
          description: New JWT secret
    RotateKeyResponse:
      type: object
      properties:
        client:
          $ref: "#/components/schemas/Client"
        api_key:
          type: string
          description: New API key
...